% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isaric4c.R
\name{isaric4c}
\alias{isaric4c}
\title{ISARIC mortality probability and deterioration risk model}
\usage{
isaric4c(
  age,
  sex,
  num_comorbidities,
  respiratory_rate,
  admission_oxygen_saturation,
  glasgow_coma_scale,
  urea,
  crp,
  nosocomial,
  radiographic_chest_infiltrates,
  receiving_oxygen,
  lymphocytes,
  func
)
}
\arguments{
\item{age}{a vector of patient's age, male (1) or female(0)}

\item{sex}{a vector of patient's gender}

\item{num_comorbidities}{number of comorbidities (Chronic cardiac disease; chronic respiratory disease (
excluding asthma); chronic renal disease (estimated glomerular filtration rate ≤30); mild-to-severe liver disease;
dementia; chronic neurological conditions; connective tissue disease; diabetes mellitus (diet, tablet or
insulin-controlled); HIV/AIDS; malignancy; clinician-defined obesity.)}

\item{respiratory_rate}{respiratory rate in breaths per minute}

\item{admission_oxygen_saturation}{peripheral oxygen saturation on room air (\%)}

\item{glasgow_coma_scale}{a vector of Glasgow coma scale score}

\item{urea}{urea concentration in mmol/L}

\item{crp}{CRP in mg/L}

\item{nosocomial}{symptom onset or first positive SARS-CoV-2 PCR more than 7 days after admission. yes as 1, no as 0}

\item{radiographic_chest_infiltrates}{radiographic chest infiltrates, yes as 1, no as 0}

\item{receiving_oxygen}{whether a patient is receiving oxygen when oxygen saturation measured}

\item{lymphocytes}{lymphocyte count (* 10^9/L)}

\item{func}{a vector of function specification: either mortality as 0 or deterioration as 1}
}
\value{
mortality probability or deterioration probability of a patient in \%
}
\description{
ISARIC mortality probability and deterioration risk model
}
\examples{
isaric4c(age = 69, sex = 'male', num_comorbidities = 1, respiratory_rate = 25,
admission_oxygen_saturation = 90, glasgow_coma_scale = 10, urea = 10, crp = 70,
func = "mortality")

isaric4c(nosocomial=1, sex=1, radiographic_chest_infiltrates=1,
receiving_oxygen=1, glasgow_coma_scale=15, age=23, respiratory_rate=50,
admission_oxygen_saturation=51,urea=21, crp=32, lymphocytes=32, func = "deterioration")
}
